Added collision check.
Collision check added. Event on collision not implemented yet.

+++ Header.hpp
@@ -2,9 +2,9 @@
 #include <cstdlib>
 #include <ctime>
 
-const double G = 0.98;	//Gravity
-const double GL = 300;	//Ground Level
-const double JH = 250;	//Jump Height
+const float G = 0.98;	//Gravity
+const float GL = 300;	//Ground Level
+const float JH = 250;	//Jump Height
 
-const int WindowWidth = 800;
-const int WindowHeight = 400;
+const float WindowWidth = 800;
+const float WindowHeight = 400;
